refactor: IllegalDistanceException 예외 발생 시 BAD_REQUEST를 반환하는 함수 추가, 정의한 예외들을 common.exception 패키리도 이동
