+++ src/main/java/nextstep/subway/common/ExceptionAdvice.java
@@ -1,8 +1,9 @@
 package nextstep.subway.common;
 
 import java.util.NoSuchElementException;
-import nextstep.subway.line.application.exception.LineNotFoundException;
-import nextstep.subway.station.application.exception.StationNotFoundException;
+import nextstep.subway.common.exception.IllegalDistanceException;
+import nextstep.subway.common.exception.LineNotFoundException;
+import nextstep.subway.common.exception.StationNotFoundException;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
@@ -28,5 +29,10 @@ public class ExceptionAdvice {
     protected ResponseEntity handleEmptyResultDataAccessException(
         EmptyResultDataAccessException e) {
         return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+    }
+
+    @ExceptionHandler(IllegalDistanceException.class)
+    protected ResponseEntity handleIllegalDistanceException(IllegalDistanceException e) {
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
     }
 }
