+++ src/main/java/nextstep/subway/common/exception/IllegalDistanceException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.common.exception;
 
 public class IllegalDistanceException extends IllegalArgumentException {
     private static final String ILLEGAL_DISTANCE_ERROR = "역 사이의 거리는 %d보다 작을 수 없습니다.";
