+++ src/main/java/nextstep/subway/common/exception/LineNotFoundException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.application.exception;
+package nextstep.subway.common.exception;
 
 import java.util.NoSuchElementException;
