+++ src/main/java/nextstep/subway/common/exception/StationNotFoundException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.station.application.exception;
+package nextstep.subway.common.exception;
 
 import java.util.NoSuchElementException;
