+++ src/main/java/nextstep/subway/line/application/LineService.java
@@ -6,8 +6,8 @@ import nextstep.subway.line.domain.LineRepository;
 import nextstep.subway.line.domain.Section;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
-import nextstep.subway.line.application.exception.LineNotFoundException;
-import nextstep.subway.station.application.exception.StationNotFoundException;
+import nextstep.subway.common.exception.LineNotFoundException;
+import nextstep.subway.common.exception.StationNotFoundException;
 import nextstep.subway.station.domain.Station;
 import nextstep.subway.station.domain.StationRepository;
 import org.springframework.stereotype.Service;
