+++ src/test/java/nextstep/subway/line/domain/DistanceTest.java
@@ -1,8 +1,8 @@
 package nextstep.subway.line.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
+import nextstep.subway.common.exception.IllegalDistanceException;
 import org.junit.jupiter.api.Test;
 
 class DistanceTest {
